requirements.txt
asgiref==3.8.1
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
coverage==6.3.2
coveralls==1.7.0
Django==4.2.16
django-appconf==1.0.6
django-browser-reload==1.15.0
django-compressor==4.5.1
django-crispy-forms==1.14.0
django-tailwind==3.6.0
docopt==0.6.2
google-auth==2.35.0
h11==0.14.0
idna==3.10
Jinja2==3.1.4
Mako==1.3.6
Markdown==3.7
MarkupSafe==3.0.2
outcome==1.3.0.post0
pdoc==12.2.0
pdoc3==0.10.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments==2.18.0
PySocks==1.7.1
python-dateutil==2.9.0
rcssmin==1.1.2
requests==2.32.3
rjsmin==1.2.2
rsa==4.9
selenium==4.5.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.4.2
trio==0.27.0
trio-websocket==0.11.1
urllib3==1.26.20
wsproto==1.2.0
pytest
factory_boy